chore: lotto 패키지 추가하여 모든 파일 이동
