--- src/main/kotlin/LottoGameApplication.kt
@@ -1,14 +1,0 @@
-import controller.LottoGameController
-import model.RandomNumbersGenerator
-import util.ExceptionHandler
-import view.ConsoleLottoGameInputView
-import view.ConsoleLottoGameOutputView
-
-fun main() {
-    LottoGameController(
-        ConsoleLottoGameInputView(),
-        ConsoleLottoGameOutputView(),
-        ExceptionHandler(),
-        RandomNumbersGenerator(),
-    ).startLottoGame()
-}
