+++ src/main/kotlin/lotto/controller/LottoGameController.kt
@@ -1,14 +1,14 @@
-package controller
+package lotto.controller
 
-import model.Count
-import model.Lotto
-import model.LottoGameResult
-import model.LottoNumber
-import model.Money
-import model.NumbersGenerator
-import util.ExceptionHandler
-import view.LottoGameInputView
-import view.LottoGameOutputView
+import lotto.model.Count
+import lotto.model.Lotto
+import lotto.model.LottoGameResult
+import lotto.model.LottoNumber
+import lotto.model.Money
+import lotto.model.NumbersGenerator
+import lotto.util.ExceptionHandler
+import lotto.view.LottoGameInputView
+import lotto.view.LottoGameOutputView
 
 class LottoGameController(
     private val lottoGameInputView: LottoGameInputView,
@@ -20,20 +20,33 @@ class LottoGameController(
         val purchaseExpense: Money = getPurchaseExpense()
         var availableCount = Count.fromMoney(purchaseExpense)
         val manualCount = getManualCount(availableCount.value)
-        val manualLotteries = lottoGameInputView.inputManualLotteryNumber(manualCount.value).map { Lotto(*it.toIntArray()) }
+        val manualLotteries = getManualLotteries(manualCount)
         availableCount -= manualCount
-        val autoLotteries =
-            List(availableCount.value) {
-                Lotto(*numbersGenerator.generate(6).toIntArray())
-            }
-        lottoGameOutputView.showLottoCount((availableCount - manualCount).value, manualCount.value)
-        lottoGameOutputView.showPurchasedLotteries(autoLotteries)
-        val entireLotteries = manualLotteries + autoLotteries
+
+        val autoLotteries = getAutoLotteries(availableCount)
+        showPurchaseResult(availableCount, manualCount, autoLotteries)
+
         val winningLotto = getWinningLotto()
         val bonusLottoNumber = createBonusLottoNumber(winningLotto)
-        val lottoGameResult = LottoGameResult(bonusLottoNumber, winningLotto, entireLotteries)
-        displayLottoResult(lottoGameResult, purchaseExpense)
+        val lottoGameResult = LottoGameResult(bonusLottoNumber, winningLotto, manualLotteries + autoLotteries)
+        showLottoGameResult(lottoGameResult, purchaseExpense)
     }
+
+    private fun showPurchaseResult(
+        availableCount: Count,
+        manualCount: Count,
+        autoLotteries: List<Lotto>,
+    ) {
+        lottoGameOutputView.showLottoCount(availableCount.value, manualCount.value)
+        lottoGameOutputView.showPurchasedLotteries(autoLotteries)
+    }
+
+    private fun getAutoLotteries(availableCount: Count) = List(availableCount.value) { Lotto(*numbersGenerator.generate().toIntArray()) }
+
+    private fun getManualLotteries(manualCount: Count) =
+        exceptionHandler.handleInputValue {
+            lottoGameInputView.inputManualLotteryNumber(manualCount.value).map { Lotto(*it.toIntArray()) }
+        }
 
     private fun getPurchaseExpense(): Money = Money(exceptionHandler.handleInputValue(lottoGameInputView::inputPurchaseExpense))
 
@@ -51,7 +64,7 @@ class LottoGameController(
             LottoNumber(bonusNumber, winningLotto)
         }
 
-    private fun displayLottoResult(
+    private fun showLottoGameResult(
         lottoGameResult: LottoGameResult,
         purchaseExpense: Money,
     ) {
