+++ src/main/kotlin/lotto/model/LottoNumber.kt
@@ -1,4 +1,4 @@
-package model
+package lotto.model
 
 @JvmInline
 value class LottoNumber(val number: Int) {
