+++ src/main/kotlin/lotto/model/Rank.kt
@@ -1,4 +1,4 @@
-package model
+package lotto.model
 
 enum class Rank(val countOfMatch: Int, val winningMoney: Int) {
     FIRST(6, 2_000_000_000),
