+++ src/test/kotlin/lotto/model/MoneyTest.kt
@@ -1,4 +1,4 @@
-package model
+package lotto.model
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
